requirements.txt
Django==2.0.3
django-environ==0.4.4
environ==1.0
psycopg2==2.6.2
djangorestframework==3.7.7
django-filter==1.1.0